generator/requirements.txt
psycopg2-binary==2.9.9
Faker==33.3.1